refactor voting results to component
