+++ apps/core-app/src/components/proposalCards/ReadyForProcessing.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import { ActionTemplate } from './ActionPrimitives';
+
+export const ReadyForProcessing = () => {
+  return <ActionTemplate statusDisplay='Ready For Processing' />
